Merge and adapt CI system from 0.2.8

Closes #11, #14, and #15

See merge request armagetronad/armagetronad!42

+++ bootstrap.sh
@@ -16,7 +16,7 @@ if test -r batch/make/version; then
     rm -f version
 fi
 echo "Copying license..."
-cp COPYING.txt COPYING
+test -r COPYING || cp COPYING.txt COPYING || exit $?
 echo "Running aclocal..."
 $ACLOCAL || { rm aclocal.m4; exit 1; }
